requirements.txt
# Autres
python-dotenv==1.0.0

# Packages principales
mysql-connector-python==8.0.33
SQLAlchemy==1.4.7
python-dotenv==1.0.0
psycopg2-binary==2.9.6
PyMySQL==1.0.3
pandas==1.4.4

# Pour API
uvicorn==0.22.0
fastapi==0.95.2

# ajouter
mysqlclient
joblib>=1.1.0
scikit-learn>=1.3.0

